requirements.txt
flask
python-dotenv
flask-wtf
flask-sqlalchemy
flask-migrate
psycopg2-binary